Add an option to use Bootstrap (with vue-strap)
It's good to have the option of having the generator set up bootstrap for you.

In addition, along the way I fixed two bugs that were getting in my way of implementing this.

1) The generated package.json files did't have a dependency on `file-loader` or `url-loader`, despite the webpack.*.js files being set up to use them for png an jpg images.
This would cause sites using this generator to crash if they tried to use a png or jpg image. The same would also apply now to ttf, eot, and svg files. (Bootstrap uses files of those types for fonts and icons.)

2) The .editorconfig file wasn't being generated properly. This was a minor, easily fixed bug, as it was due to the template being mistakingly named `editorconfig` ant not `_editorconfig`.